+++ domain/src/test/kotlin/com/mickeycj/domain/usecases/GetCountriesUseCaseSpec.kt
@@ -8,13 +8,13 @@ import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
 
-import com.mickeycj.domain.Faker
+import com.mickeycj.domain.utils.Faker
 import com.mickeycj.domain.contracts.CountryRepository
 
 /**
  * Spek tests for getting countries from the repository.
  */
-object GetCountriesUseCaseSpek : Spek({
+object GetCountriesUseCaseSpec : Spek({
 
     describe("Get Countries Use Case") {
