+++ domain/src/test/kotlin/com/mickeycj/domain/utils/Faker.kt
@@ -1,4 +1,4 @@
-package com.mickeycj.domain
+package com.mickeycj.domain.utils
 
 import io.reactivex.Single
